% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_armour}
\alias{set_armour}
\title{Set Armour (war gear)}
\usage{
set_armour(fields, armour, protection, load)
}
\arguments{
\item{fields}{list of pdf fields}

\item{armour}{Name of armour}

\item{protection}{Protection value}

\item{load}{Load value}
}
\description{
Set Armour (war gear)
}
