% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_calling}
\alias{set_calling}
\title{Set Character Calling}
\usage{
set_calling(fields, calling, shadow_path)
}
\arguments{
\item{fields}{list of pdf fields}

\item{calling}{Calling name}

\item{shadow_path}{Calling shadow path}
}
\description{
Set Character Calling
}
