% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_conditions}
\alias{set_conditions}
\title{Set Character Conditions}
\usage{
set_conditions(
  fields,
  is_weary = FALSE,
  is_miserable = FALSE,
  is_wounded = FALSE,
  injury = ""
)
}
\arguments{
\item{fields}{list of pdf fields}

\item{is_weary}{Boolean}

\item{is_miserable}{Boolean}

\item{is_wounded}{Boolean}

\item{injury}{Injury description}
}
\description{
Set Character Conditions
}
