% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_culture}
\alias{set_culture}
\title{Set Character Culture}
\usage{
set_culture(fields, culture, blessing)
}
\arguments{
\item{fields}{list of pdf fields}

\item{culture}{Culture name}

\item{blessing}{Culture blessing}
}
\description{
Set Character Culture
}
