% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_current_endurance}
\alias{set_current_endurance}
\title{Set Current Endurance}
\usage{
set_current_endurance(fields, endurance, load, fatigue)
}
\arguments{
\item{fields}{list of pdf fields}

\item{endurance}{Current endurance}

\item{load}{Current load value}

\item{fatigue}{Character fatigue}
}
\description{
Set Current Endurance
}
