% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_current_hope}
\alias{set_current_hope}
\title{Set Current Hope}
\usage{
set_current_hope(fields, hope, shadow, shadow_scars)
}
\arguments{
\item{fields}{list of pdf fields}

\item{hope}{Current hope}

\item{shadow}{Current Shadow}

\item{shadow_scars}{Character Shadow Scars}
}
\description{
Set Current Hope
}
