% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_features}
\alias{set_features}
\title{Set Character Features}
\usage{
set_features(fields, features)
}
\arguments{
\item{fields}{list of pdf fields}

\item{features}{vector of character features}
}
\description{
Set Character Features
}
