% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_flaws}
\alias{set_flaws}
\title{Set Character Flaws}
\usage{
set_flaws(fields, flaws)
}
\arguments{
\item{fields}{list of pdf fields}

\item{flaws}{vector of character flaws}
}
\description{
Set Character Flaws
}
