% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_heart}
\alias{set_heart}
\title{Set Character Heart attribute}
\usage{
set_heart(fields, rating, target, hope)
}
\arguments{
\item{fields}{list of pdf fields}

\item{rating}{Heart Rating}

\item{target}{Heart Target Number}

\item{hope}{Hope Rating}
}
\description{
Set Character Heart attribute
}
