% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_helm}
\alias{set_helm}
\title{Set Helm (war gear)}
\usage{
set_helm(fields, helm, protection, load)
}
\arguments{
\item{fields}{list of pdf fields}

\item{helm}{Name of helm}

\item{protection}{Protection value}

\item{load}{Load value}
}
\description{
Set Helm (war gear)
}
