% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_patron}
\alias{set_patron}
\title{Set Fellowship Patron}
\usage{
set_patron(fields, patron)
}
\arguments{
\item{fields}{list of pdf fields}

\item{patron}{Patron name}
}
\description{
Set Fellowship Patron
}
