% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_points}
\alias{set_points}
\title{Set Various Character Points}
\usage{
set_points(fields, adventure, skill, fellowship)
}
\arguments{
\item{fields}{list of pdf fields}

\item{adventure}{Adventure Points}

\item{skill}{Skill Points}

\item{fellowship}{Fellowship Points}
}
\description{
Set Various Character Points
}
