% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_proficiency_level}
\alias{set_proficiency_level}
\title{Set Combat Proficiency}
\usage{
set_proficiency_level(fields, proficiency, lvl)
}
\arguments{
\item{fields}{list of pdf fields}

\item{proficiency}{Name of Proficiency}

\item{lvl}{Proficiency level (integer)}
}
\description{
Set Combat Proficiency
}
