% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_shield}
\alias{set_shield}
\title{Set Shield (war gear)}
\usage{
set_shield(fields, shield, parry, load)
}
\arguments{
\item{fields}{list of pdf fields}

\item{shield}{Name of shield}

\item{parry}{Parry value}

\item{load}{Load value}
}
\description{
Set Shield (war gear)
}
