% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_skill_level}
\alias{set_skill_level}
\title{Set skill level}
\usage{
set_skill_level(fields, skill, lvl)
}
\arguments{
\item{fields}{list of pdf fields}

\item{skill}{Name of skill}

\item{lvl}{Skill level (integer)}
}
\description{
Set skill level
}
