% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_skills_favored}
\alias{set_skills_favored}
\title{Set favored skills}
\usage{
set_skills_favored(fields, skills)
}
\arguments{
\item{fields}{list of pdf fields}

\item{skills}{vector of character skills}
}
\description{
Set favored skills
}
