% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_standard_of_living}
\alias{set_standard_of_living}
\title{Set Character Standard of Living}
\usage{
set_standard_of_living(fields, living, treasure)
}
\arguments{
\item{fields}{list of pdf fields}

\item{living}{Standard of Living}

\item{treasure}{Initial treasure value}
}
\description{
Set Character Standard of Living
}
