% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_strength}
\alias{set_strength}
\title{Set Character Strength attribute}
\usage{
set_strength(fields, rating, target, endurance)
}
\arguments{
\item{fields}{list of pdf fields}

\item{rating}{Strength Rating}

\item{target}{Strength Target Number}

\item{endurance}{Endurance Rating}
}
\description{
Set Character Strength attribute
}
