% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_travelling_gear}
\alias{set_travelling_gear}
\title{Set Travelling Gear}
\usage{
set_travelling_gear(fields, gear)
}
\arguments{
\item{fields}{list of pdf fields}

\item{gear}{vector of gear descriptions}
}
\description{
Set Travelling Gear
}
