% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_valour}
\alias{set_valour}
\title{Set Character Valour and Rewards}
\usage{
set_valour(fields, valour, rewards = "")
}
\arguments{
\item{fields}{list of pdf fields}

\item{valour}{Valour value (integer)}

\item{rewards}{vector of Rewards descriptions}
}
\description{
Set Character Valour and Rewards
}
