% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_weapons}
\alias{set_weapons}
\title{Set weapons (war gear)}
\usage{
set_weapons(fields, weapon, damage, injury, load, notes)
}
\arguments{
\item{fields}{list of pdf fields}

\item{weapon}{Name of weapon}

\item{damage}{weapon damage}

\item{injury}{weapon injury target number}

\item{load}{load value}

\item{notes}{notes}
}
\description{
Set weapons (war gear)
}
