% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_wisdom}
\alias{set_wisdom}
\title{Set Character Wisdom and Virtues}
\usage{
set_wisdom(fields, wisdom, virtues = "")
}
\arguments{
\item{fields}{list of pdf fields}

\item{wisdom}{Wisdom value (integer)}

\item{virtues}{vector of Virtues descriptions}
}
\description{
Set Character Wisdom and Virtues
}
