% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.r
\name{set_wits}
\alias{set_wits}
\title{Set Character Wits attribute}
\usage{
set_wits(fields, rating, target, parry)
}
\arguments{
\item{fields}{list of pdf fields}

\item{rating}{Wits Rating}

\item{target}{Wits Target Number}

\item{parry}{Parry Rating}
}
\description{
Set Character Wits attribute
}
