% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.r
\name{write_pdf}
\alias{write_pdf}
\title{Write pdf fields}
\usage{
write_pdf(fields, outputfile, template = c("men", "dwarf", "elf", "hobbit"))
}
\arguments{
\item{fields}{character data}

\item{outputfile}{name of output file}

\item{template}{pdf template type}
}
\description{
Write pdf fields
}
